[profile] Fix profile merging with binary IDs
This fixes support for merging profiles which broke as a consequence
of 2aec07e121665e449317c3a4d34cfecf1cb43c16. The issue was missing
adjustment in merge logic to account for the binary IDs which are
now included in the raw profile just after header.

In addition, this change also:
* Includes the version in module signature that's used for merging
to avoid accidental attempts to merge incompatible profiles.
* Moves the binary IDs size field after version field in the header
as was suggested in the review.

Differential Revision: https://reviews.llvm.org/D107143