[InstCombine] factorize min/max intrinsic ops with common operand (2nd try)
This is a re-try of de40231fa068 which was reverted because
it missed a null check. Extra test for that failure added.

Original commit message:
This is an adaptation of D41603 and another step on the way
to canonicalizing to the intrinsic forms of min/max.

See D98152 for status.